- [ZBXNEXT-636] updated to the latest trunk (Aly)
git-svn-id: svn://svn.zabbix.com/branches/dev@17628 f4e7fd02-d210-4e34-a1e0-d4093fb66cc2

+++ ZBXNEXT-636/frontends/php/api/classes/class.cdiscoveryrule.php
@@ -1,7 +1,7 @@
 <?php
 /*
 ** ZABBIX
-** Copyright (C) 2000-2010 SIA Zabbix
+** Copyright (C) 2000-2011 SIA Zabbix
 **
 ** This program is free software; you can redistribute it and/or modify
 ** it under the terms of the GNU General Public License as published by
@@ -974,9 +974,10 @@ COpt::memoryPick();
 			while($item = DBfetch($db_items)){
 				$iprototypeids[$item['itemid']] = $item['itemid'];
 			}
-			if(!CItemPrototype::delete($iprototypeids, true))
-				self::exception(ZBX_API_ERROR_PARAMETERS, 'Cannot delete discovery rule');
-
+			if(!empty($iprototypeids)){
+				if(!CItemPrototype::delete($iprototypeids, true))
+					self::exception(ZBX_API_ERROR_PARAMETERS, _('Cannot delete discovery rule'));
+			}
 
 			DB::delete('items', array('itemid'=>$ruleids));
